my_gemini_agent/pyproject.toml
[tool.poetry]
name = "my-gemini-agent"
version = "0.1.0"
description = "A Gemini agent built with ADK"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
google-generativeai = "^0.3.0"
google-adk = "^0.1.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
my-gemini-agent = "src.__main__:main"
